refactoring, sanitize whole body
