+++ lib/auth.js
@@ -432,6 +432,19 @@ function verifyRole(...roles) {
   };
 }
 
+function requireRoles(condition, ...roles) {
+  return function(req, res, next) {
+    let pre = true;
+    if (_.isFunction(condition)) {
+      pre = condition(req, res);
+    }
+    if (!pre) {
+      return next();
+    }
+    return verifyRole(...roles)(req, res, next);
+  };
+}
+
 var basic = require('basic-auth');
 
 function notKnown(cred) {
@@ -455,6 +468,7 @@ function basicAuth(req, res, next) {
 module.exports = {
   ensureAuthenticated: ensureAuthenticated,
   verifyRole: verifyRole,
+  requireRoles: requireRoles,
   checkAuth: checkAuth,
   sessionLocals: sessionLocals,
   basicAuth: basicAuth,
