+++ lib/req-utils.js
@@ -1,6 +1,5 @@
 var sanitize = require('google-caja-sanitizer').sanitize;
 var _ = require('lodash');
-var auth = require('../lib/auth');
 
 /**
  * Check the property list of http request. Set the property to null if it is
@@ -40,7 +39,7 @@ var sanitizeValue = function(val) {
     return null;
   }
   if (_.isObject(val)) {
-    return _.mapObject(val, function(input) {
+    return _.mapValues(val, function(input) {
       return sanitizeValue(input);
     });
   }
@@ -62,11 +61,15 @@ var sanitizeValue = function(val) {
  */
 function sanitizeMw(list, names) {
   return function(req, res, next) {
-    names.forEach(function(n) {
-      if (req[list].hasOwnProperty(n)) {
-        req[list][n] = sanitizeValue(req[list][n]);
-      }
-    });
+    if (names.length === 0) {
+      req[list] = sanitizeValue(req[list]);
+    } else {
+      names.forEach(function(n) {
+        if (req[list].hasOwnProperty(n)) {
+          req[list][n] = sanitizeValue(req[list][n]);
+        }
+      });
+    }
     next();
   };
 }
@@ -320,19 +323,6 @@ function getSharedGroup(sharedGroup, id) {
   return -1;
 }
 
-function requireRoles(condition, ...roles) {
-  return function(req, res, next) {
-    let pre = true;
-    if (_.isFunction(condition)) {
-      pre = condition(req, res);
-    }
-    if (!pre) {
-      return next();
-    }
-    return auth.verifyRole(...roles)(req, res, next);
-  };
-}
-
 module.exports = {
   filter: filter,
   hasAll: hasAll,
@@ -350,6 +340,5 @@ module.exports = {
   getAccess: getAccess,
   getSharedWith: getSharedWith,
   getSharedGroup: getSharedGroup,
-  requireRoles: requireRoles,
   requireOwner: requireOwner,
 };
